feat: @types renamed
