+++ src/App/Services/Observers/MatchService.ts
@@ -1,4 +1,4 @@
-import { IObserver, ISubject } from '../../../@types/IObserver'
+import { IObserver, ISubject } from '../../../types/IObserver'
 import { SummonerV4DTO } from '@jlenon7/zedjs/dist/models-dto'
 
 class MatchService implements IObserver {
