+++ src/Database/factory.ts
@@ -1,9 +1,9 @@
 import faker from 'faker'
 import { getRepository } from 'typeorm'
 
-import IFactory from '../@types/IFactory'
-import ISummoner, { ISummonerObject } from '../@types/ISummoner'
-import IChampion from '../@types/IChampion'
+import IFactory from '../types/IFactory'
+import ISummoner, { ISummonerObject } from '../types/ISummoner'
+import IChampion from '../types/IChampion'
 import Summoner from '../App/Models/Summoner'
 import Champion from '../App/Models/Champion'
