+++ src/types/IObserver.ts
